==> notebooks/1.0-sk-extract-titanic-data.ipynb <==
### Extracting Titanic Disaster Data from Kaggle

from dotenv import load_dotenv, find_dotenv

#find .env automatically by walking up directories until it is found
dotenv_path = find_dotenv()
#load up the entries as environment variables
load_dotenv(dotenv_path)

#extracting environment variable using os.environ.get
import os
KAGGLE_USERNAME = os.environ.get("KAGGLE_USERNAME")
print(KAGGLE_USERNAME)

#imports
import requests
from requests import session
import os
from dotenv import load_dotenv, find_dotenv

#payload for post
payload = {
    'action': 'login',
    'username': os.environ.get("KAGGLE_USERNAME"),
    'password': os.environ.get("KAGGLE_PASSWORD")
}

#URL for train file (getting the link from KAGGLE Website)
url1 = 'https://www.kaggle.com/c/titanic/download/train.csv'

#Setup session
with session() as c:
    #post request
    c.post('https://www.kaggle.com/account/login', data = payload)
    #get request
    response = c.get(url1)
    #print response text
    print(response.text)

from requests import session
#payload 
payload = {
    'action': 'login',
    'username': os.environ.get("KAGGLE_USERNAME"),
    'password': os.environ.get("KAGGLE_PASSWORD")
}

def extract_data(url,file_path):
    '''
    extract data from kaggle'''
    
    #setup session
    with session() as c:
        c.post('https://www.kaggle.com/account/login', data = payload)
        #open files to write
        with open(file_path, 'w') as handle:
            response = c.get(url, stream=True)
            for block in response.iter_content(1024):
                handle.write(block)

#urls
train_url = 'https://www.kaggle.com/c/titanic/download/train.csv'
test_url = 'https://www.kaggle.com/c/titanic/download/test.csv'

#file_paths
raw_data_path = os.path.join(os.path.pardir,'data','raw')
train_data_path = os.path.join(raw_data_path, 'train.csv')
test_data_path = os.path.join(raw_data_path, 'test.csv')

#extract data
extract_data(train_url,train_data_path)
extract_data(test_url,test_data_path)

!ls -1 ../data/raw

### Building the file script

get_raw_data_script_file = os.path.join(os.path.pardir, 'src', 'data', 'get_raw_data.py')

%%writefile $get_raw_data_script_file
# -*- coding: utf-8 -*-
import os
from dotenv import load_dotenv, find_dotenv
from requests import session
import logging

#payload for login into Kaggle
payload = {
    'action': 'login',
    'username': os.environ.get("KAGGLE_USERNAME"),
    'password': os.environ.get("KAGGLE_PASSWORD")
}

def extract_data(url,file_path):
    '''
    method to extract data'''
    
    #setup session
    with session() as c:
        c.post('https://www.kaggle.com/account/login', data = payload)
        #open files to write
        with open(file_path, 'w') as handle:
            response = c.get(url, stream=True)
            for block in response.iter_content(1024):
                handle.write(block)
                
def main(project_dir):
    '''
    main method'''
    #get logger
    logger = logging.getLogger(__name__)
    logger.info('getting raw data')
    #urls
    train_url = 'https://www.kaggle.com/c/titanic/download/train.csv'
    test_url = 'https://www.kaggle.com/c/titanic/download/test.csv'

    #file_paths
    raw_data_path = os.path.join(os.path.pardir,'data','raw')
    train_data_path = os.path.join(raw_data_path, 'train.csv')
    test_data_path = os.path.join(raw_data_path, 'test.csv')
    
    #extract data
    extract_data(train_url,train_data_path)
    extract_data(test_url,test_data_path)
    logger.info('download raw training and test data')
    
if __name__ == '__main__':
    #getting root directory
    project_dir = os.path.join(os.path.dirname(__file__), os.pardir, os.pardir)
    
    #setup logger
    log_fmt = '%(asctime)s - %(name)s - %(levelname)s - %(message)s'
    logging.basicConfig(level=logging.INFO, format = log_fmt)
    
    #find .env automatically by walking up directories until it is found
    dotenv_path = find_dotenv()
    #load up the entries as environment variables
    load_dotenv(dotenv_path)
    
    #call the main
    main(project_dir)

!python $get_raw_data_script_file
