==> notebooks/3.0-sk-building-predicitve-models.ipynb <==
## Building Predictive Models

import pandas as pd
import os
import numpy as np

### Import data

#set the path for the processed data
processed_data_path = os.path.join(os.path.pardir,'data','processed')
train_file_path = os.path.join(processed_data_path, 'train.csv')
test_file_path = os.path.join(processed_data_path, 'test.csv')

#read the data with all default parameters
train_df = pd.read_csv(train_file_path, index_col='PassengerId')
test_df = pd.read_csv(test_file_path, index_col='PassengerId')

train_df.info()

test_df.info()

### Data Preparation

X = train_df.loc[:,'Age':].as_matrix().astype('float')
y = train_df['Survived'].ravel()

print X.shape, y.shape

# train test split
from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(X,y,test_size = 0.2, random_state = 0)
print X_train.shape, y_train.shape
print X_test.shape, y_test.shape

#it is advisable to check the number of positive class values in the initial data model
#average survival rate in train and test
print 'mean survival in train :  {0:.3f}'.format(np.mean(y_train))
print 'mean survival in test :  {0:.3f}'.format(np.mean(y_test))

### Check Scikit-learn version

import sklearn
sklearn.__version__

#!conda update -y scikit-learn

### Baseline Model

#import function
from sklearn.dummy import DummyClassifier

#create model
model_dummy = DummyClassifier(strategy = 'most_frequent', random_state=0)

#train model
model_dummy.fit(X_train,y_train)

print 'score for baseline model : {0:.2f}'.format(model_dummy.score(X_test,y_test))

#performance metrics
from sklearn.metrics import accuracy_score, confusion_matrix, precision_score, recall_score

#accuracy score
print 'accuracy score for baseline model: {0:.2f}'.format(accuracy_score(y_test, model_dummy.predict(X_test)))

#confusion matrix
print 'confusion matrix for baseline model: \n {0}'.format(confusion_matrix(y_test, model_dummy.predict(X_test)))

#precision and recall scores
print 'precision score for baseline model: {0:.2f}'.format(precision_score(y_test, model_dummy.predict(X_test)))
print 'recall score for baseline model: {0:.2f}'.format(recall_score(y_test, model_dummy.predict(X_test)))

### First Kaggle Submission

#converting to the matrix
test_X = test_df.as_matrix().astype('float')

#get predictions
predictions = model_dummy.predict(test_X)

df_submissions = pd.DataFrame({'PassengerId': test_df.index, 'Survived': predictions})

df_submissions.head()

submissions_data_path = os.path.join(os.path.pardir, 'data', 'external')
submissions_file_path = os.path.join(submissions_data_path, '01_dummy.csv')

df_submissions.to_csv(submissions_file_path, index=False)

def get_submission_file(model, filename):
    #converting to the matrix
    test_X = test_df.as_matrix().astype('float')
    #make predictions
    predictions = model_dummy.predict(test_X)
    #submissions dataframe
    df_submissions = pd.DataFrame({'PassengerId': test_df.index, 'Survived': predictions})
    #submissions file
    submissions_data_path = os.path.join(os.path.pardir, 'data', 'external')
    submissions_file_path = os.path.join(submissions_data_path, filename)
    #write to the file
    df_submissions.to_csv(submissions_file_path, index=False)
    

#get submissions file
get_submission_file(model_dummy,'01_dummy.csv')

### Logistic Regression Model

#import function
from sklearn.linear_model import LogisticRegression

#create model
model_lr_1 = LogisticRegression(random_state = 0)

#train model
model_lr_1.fit(X_train,y_train)

#evaluation model
print 'score for logistic regression - version 1: {0:.2f}'.format(model_lr_1.score(X_test,y_test))

# performance metrics
#accuracy score
print 'accuracy score for logistic regression - version 1: {0:.2f}'.format(accuracy_score(y_test, model_lr_1.predict(X_test)))
#confusion matrix
print 'confusion matrix for logistic regression - version 1: \n{0}'.format(confusion_matrix(y_test, model_lr_1.predict(X_test)))
#precision score
print 'precision score for logistic regression - version 1: {0:.2f}'.format(precision_score(y_test, model_lr_1.predict(X_test)))
#recall score
print 'recall score for logistic regression - version 1: {0:.2f}'.format(recall_score(y_test, model_lr_1.predict(X_test)))

#model coefficients (Model Weights or Model Parameters)
model_lr_1.coef_

### Second Kaggle Parameters

#get submission file
get_submission_file(model_lr_1,'02_lr.csv')

test_X = test_df.as_matrix().astype('float')

predictions = model_lr_1.predict(test_X)

df_submissions = pd.DataFrame({'PassengerId': test_df.index, 'Survived': predictions})

submissions_data_path = os.path.join(os.path.pardir, 'data', 'external')
submissions_file_path = os.path.join(submissions_data_path, '02_lr.csv')

df_submissions.to_csv(submissions_file_path, index=False)

## Hyperparameter Optimization

#base model
model_lr = LogisticRegression(random_state = 0)

from sklearn.model_selection import GridSearchCV

parameters = {'C': [1.0, 10.0, 50.0, 100.0, 1000.0], 'penalty': ['l1','l2']}
clf = GridSearchCV(model_lr, param_grid=parameters, cv=3)

clf.fit(X_train,y_train)

clf.best_params_

print 'best score : {0:.2f}'.format(clf.best_score_)

#evaluate the model
print 'score for logistic regression - version 2 : {0:.2f}'.format(clf.score(X_test, y_test))

### Making third submission

#get submission file
get_submission_file(clf,'03_lr.csv')

test_X = test_df.as_matrix().astype('float')

predictions = clf.predict(test_X)

df_submissions = pd.DataFrame({'PassengerId': test_df.index, 'Survived': predictions})

submissions_data_path = os.path.join(os.path.pardir, 'data', 'external')
submissions_file_path = os.path.join(submissions_data_path, '03_lr.csv')

df_submissions.to_csv(submissions_file_path, index=False)

### Feature Normalization and Standardization

from sklearn.preprocessing import MinMaxScaler, StandardScaler

#### Feature Normalization

#feature normalization
scaler = MinMaxScaler()
X_train_scaled = scaler.fit_transform(X_train)

X_train_scaled[:,0].min(),X_train_scaled[:,0].max()

#normalize the test data as well
X_test_scaled = scaler.transform(X_test)

#### Feature Standardization

#feature standardization
scaler = StandardScaler()
X_train_scaled = scaler.fit_transform(X_train)
X_test_scaled = scaler.transform(X_test)

#### Create model after standardization

#base model
model_lr = LogisticRegression()
parameters = {'C': [1.0, 10.0, 50.0, 100.0, 1000.0], 'penalty': ['l1','l2']}
clf = GridSearchCV(model_lr, param_grid=parameters, cv=3)
clf.fit(X_train_scaled,y_train)

clf.best_score_

#evaluate the model
print 'score for logistic regression - version 2 : {0:.2f}'.format(clf.score(X_test_scaled, y_test))

## Model Persistence

#import pickle library
import pickle

#create the file paths
model_file_path = os.path.join(os.path.pardir, 'models', 'lr_model.pkl')
scaler_file_path = os.path.join(os.path.pardir, 'models', 'lr_scaler.pkl')

#open the files to write
model_file_pickle = open(model_file_path, 'wb')
scaler_file_pickle = open(scaler_file_path, 'wb')

#persist the model and the scaler
pickle.dump(clf, model_file_pickle)
pickle.dump(scaler, scaler_file_pickle)

#close the files
model_file_pickle.close()
scaler_file_pickle.close()

#### Load the persisted file

#open files in read mode
model_file_pickle = open(model_file_path, 'r')
scaler_file_pickle = open(scaler_file_path, 'r')
#load files
clf_loaded = pickle.load(model_file_pickle)
scaler_loaded = pickle.load(scaler_file_pickle)
#close the files
model_file_pickle.close()
scaler_file_pickle.close()

clf_loaded

scaler_loaded

#transform the test data using loaded scaler object
X_test_scaled = scaler_loaded.transform(X_test)
#calculate the score using the loaded scaler object
print 'score for logistic regression - version 2 : {0:.2f}'.format(clf_loaded.score(X_test_scaled, y_test))
